Initial commit.
We're using Java for this project cause the library that converts MathML
to LaTeX is in Java. There's a lot of libraries that can convert LaTeX
to MathML, but not the other around. BBL exported the assessment with
the solution formula in MathML format and Webwork expects TeX, so we
need a conversion library.