Parsing formula variables.
The formula variables are now stored in the questions. It turns out that
you can configure the number of significant digits generated by the
random range. They labelled it as "scale" and I've just called them
significant digits. Unfortunately, I'm not sure if the range is
inclusive or exclusive and not sure how the sig digits works when it's 0.
There's instances where sig digits is set to 1 but you're given a range
from 10 to 25 which obviously is problematic if they really want only 1
sig digit.

Maybe they meant decimal places? But on the UI side, they said sig
digits...