Move CalculatedQuestion methods out from Question.
Since we started with only calculated questions, it made sense to have
only 1 question class. But now that we have multiple question types,
there needs to be a separate question classes, so now we have
CalculatedQuestion and MultiChoiceQuestion classes.

This commit just moves the Calculated Question specific methods and
variables to the CalculatedQuestion class.