Question attempt messages and randomize choices.
When a student attempt a problem, BBL Test lets you set a message for
the student if they got the question correct or if they get the question
wrong. Webwork seems to have something equivalent, at least in terms of
problem hints and possibly congratulation messages (might need
achievements enabled). So now we have parsing for these messages.

Choices in a MultiChoiceQuestion may be presented in a random order or
in a set order. Webwork allows for this, so a new property for
MultiChoiceQuestion now stores whether the choices should be randomized.