Fix exception when no correct/incorrect messages.
Messages to the student when they get an answer correct or incorrect
sometimes may not be set. The converter throws an exception cause the
parser was expecting it to be set. Fixed to ignore unset
correct/incorrect messages by checking that the child elements that
describe the messages actually exist.